course_project/requirements.txt
pandas>=2.0.0
requests>=2.31.0
matplotlib>=3.8.0
seaborn>=0.13.0
statsmodels>=0.14.0
numpy>=1.25.0
snakemake>=7.32.0
python-dotenv>=1.0.0
